+++ example/index.tsx
@@ -1,7 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
-import configKeepAlive from 'keep-alive-comp/dist/configKeepAlive';
+// import { configKeepAlive } from 'keep-alive-comp';
+
+// console.log(configKeepAlive);
 
 // optional/可以不写
 // configKeepAlive({ maxLength: 2, useStorage: 'sessionStorage' });